sample JV data
Sample data of a number of hole-only devices made in our lab to be used to learn to operate the FitSCLC program. In principle, the devices are measured up to high enough voltages to reach the space charge limited conduction (SCLC) regime, but this is not always so. Likewise, some samples show trap-related phenomena at lower voltages and/or series resistance limitations at higher voltages. The build-in voltage of the devices is typically small (<0.3 V). Have fun!